Extract Users query helper
Prior to this change, the query helper for users query was inside
of UsersFilter, but it was also needed on other places.

This change extracts the Helper component for reuseablitity.
